Edit app + Delete app, OK

+++ src/main/java/ch/heigvd/amt/projet/services/ApplicationDAO.java
@@ -49,7 +49,12 @@ public class ApplicationDAO extends DatabaseUtils implements ApplicationDaoLocal
 
     @Override
     public boolean deleteApp(Application app) {
-        String sql = "SELECT APIToken FROM applications WHERE APIToken = ?;";
+        return deleteApp(app.getAPI_TOKEN(), app.getAppOwner());
+    }
+
+    @Override
+    public boolean deleteApp(String APIToken, String appOwner) {
+        String sql = "SELECT APIToken FROM applications WHERE APIToken = ? and appOwner = ?;";
         ResultSet resultSet = null;
         boolean result = false;
         PreparedStatement preparedStatement    = null;
@@ -58,7 +63,8 @@ public class ApplicationDAO extends DatabaseUtils implements ApplicationDaoLocal
         try (Connection connection = dataSource.getConnection()) {
             preparedStatement = connection.prepareStatement(sql);
 
-            preparedStatement.setString(1, app.getAPI_TOKEN());
+            preparedStatement.setString(1, APIToken);
+            preparedStatement.setString(2, appOwner);
             resultSet = preparedStatement.executeQuery();
 
             if(!resultSet.next()) {
@@ -66,9 +72,10 @@ public class ApplicationDAO extends DatabaseUtils implements ApplicationDaoLocal
             }
             else{
                 resultSet.beforeFirst();
-                String sqlDel = "DELETE FROM applications WHERE APIToken = ?;";
+                String sqlDel = "DELETE FROM applications WHERE APIToken = ? and appOwner = ?;";
                 preparedStatementDel = connection.prepareStatement(sqlDel);
-                preparedStatementDel.setString(1, app.getAPI_TOKEN());
+                preparedStatementDel.setString(1, APIToken);
+                preparedStatementDel.setString(2, appOwner);
                 preparedStatementDel.execute();
                 return true;
             }
@@ -81,11 +88,16 @@ public class ApplicationDAO extends DatabaseUtils implements ApplicationDaoLocal
     }
 
     @Override
-    public boolean updateApp(Application app, String newName, String newDescription) {
-        if(newName == null || newDescription == null){
+    public boolean updateApp(Application app, String newName, String newDescription, String appOwner) {
+        return updateApp(app.getAPI_TOKEN(), newName, newDescription, appOwner);
+    }
+
+    @Override
+    public boolean updateApp(String APIToken, String newName, String newDescription, String appOwner) {
+        if(APIToken == null || newName == null || newDescription == null){
             return false;
         }
-        String sql = "SELECT APIToken FROM applications WHERE APIToken = ?;";
+        String sql = "SELECT APIToken FROM applications WHERE APIToken = ? and appOwner = ?;";
         ResultSet resultSet = null;
         boolean result = false;
         PreparedStatement preparedStatement    = null;
@@ -94,7 +106,8 @@ public class ApplicationDAO extends DatabaseUtils implements ApplicationDaoLocal
         try (Connection connection = dataSource.getConnection()) {
             preparedStatement = connection.prepareStatement(sql);
 
-            preparedStatement.setString(1, app.getAPI_TOKEN());
+            preparedStatement.setString(1, APIToken);
+            preparedStatement.setString(2, appOwner);
             resultSet = preparedStatement.executeQuery();
 
             if(!resultSet.next()) {
@@ -102,11 +115,12 @@ public class ApplicationDAO extends DatabaseUtils implements ApplicationDaoLocal
             }
             else{
                 resultSet.beforeFirst();
-                String sqlUpdate = "UPDATE applications SET appName = ?, description = ? WHERE APIToken = ?;";
+                String sqlUpdate = "UPDATE applications SET appName = ?, description = ? WHERE APIToken = ? and appOwner = ?;";
                 preparedStatementDel = connection.prepareStatement(sqlUpdate);
                 preparedStatementDel.setString(1, newName);
                 preparedStatementDel.setString(2, newDescription);
-                preparedStatementDel.setString(3, app.getAPI_TOKEN());
+                preparedStatementDel.setString(3, APIToken);
+                preparedStatementDel.setString(4, appOwner);
                 preparedStatementDel.execute();
                 return true;
             }
@@ -116,7 +130,6 @@ public class ApplicationDAO extends DatabaseUtils implements ApplicationDaoLocal
         } finally {
             cleanUp(preparedStatement, preparedStatementDel);
         }
-
     }
 
     @Override
